feat: adding new init command structure with example subcommand
