+++ tests/init/test_init.py
@@ -78,7 +78,7 @@ def make_output_scrubber(*extra_scrubbers: Callable[[str], str], **extra_tokens:
 def which_mock(mocker: MockerFixture) -> WhichMock:
     which_mock = WhichMock()
     which_mock.add("git")
-    mocker.patch("algokit.cli.init.shutil.which").side_effect = which_mock.which
+    mocker.patch("algokit.cli.init.command.shutil.which").side_effect = which_mock.which
     return which_mock
 
 
@@ -97,12 +97,13 @@ class ExtendedTemplateKey(str, Enum):
 # Define a fixture to monkeypatch TemplateKey with ExtendedTemplateKey
 @pytest.fixture(autouse=True)
 def _set_mocked_template_keys(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr("algokit.cli.init.TemplateKey", ExtendedTemplateKey)
+    monkeypatch.setattr("algokit.cli.init.helpers.TemplateKey", ExtendedTemplateKey)
 
 
 @pytest.fixture(autouse=True)
 def _set_blessed_templates(mocker: MockerFixture) -> None:
-    from algokit.cli.init import BlessedTemplateSource, init_command
+    from algokit.cli.init.__init__ import init_group as init_command
+    from algokit.cli.init.helpers import BlessedTemplateSource
 
     blessed_templates = {
         ExtendedTemplateKey.SIMPLE: BlessedTemplateSource(
@@ -135,7 +136,7 @@ def _set_blessed_templates(mocker: MockerFixture) -> None:
     (template_param,) = (p for p in init_command.params if p.name == "template_name")
     template_param.type = click.Choice(list(blessed_templates))
 
-    mocker.patch("algokit.cli.init._get_blessed_templates").return_value = blessed_templates
+    mocker.patch("algokit.cli.init.helpers._get_blessed_templates").return_value = blessed_templates
 
 
 @pytest.fixture(autouse=True)
@@ -143,7 +144,7 @@ def _override_bootstrap(mocker: MockerFixture) -> None:
     def bootstrap_mock(p: Path, *, ci_mode: bool, max_depth: int = 1) -> None:  # noqa: ARG001
         click.echo(f"Executed `algokit project bootstrap all` in {p}")
 
-    mocker.patch("algokit.cli.init.bootstrap_any_including_subdirs").side_effect = bootstrap_mock
+    mocker.patch("algokit.cli.init.command.bootstrap_any_including_subdirs").side_effect = bootstrap_mock
 
 
 def test_init_help() -> None:
